pyproject.toml
[project]
name = "todo-app-with-tkinter"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.2.0",
    "pydantic",
    "passlib",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
]
